+++ src/pages/MainPage/SkipReloadButton/index.tsx
